pyproject.toml
[project]
name = "telegram"
version = "0.1.0"
description = ""
authors = [
    {name = "sergio-r1",email = "sergio-r1@outlook.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pyrogram (>=2.0.106,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "tgcrypto (>=1.2.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "telegram" }
]
